Split out main logic from __init__.py into plugin.py, added testing framework (not yet completed)
